add continuous integration
This uses GitHub actions from the standard https://github.com/julia-actions organization to run the tests on submitted pull requests. Additionally this adds dependabot to keep these Actions up-to-date. This won't happen often so I set it to run monthly.

This also adds Aqua as a test-only dependency to run some extra checks, like if the compatibility of our dependencies is declared, which I added here.